requirements.txt
asgiref==3.6.0
dj-database-url==0.5.0
Django==3.2.14
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
whitenoise==6.2.0
pillow==9.5.0
